Add directory data/ for templates with a special syntax containing textual data to be displayed in pages.  Each of these template files is divided into sections beginning with the following line, with an optional title:
The semantic associations page has been converted into this format, and the
descriptions have been moved into a file in this format with each section
being a different glyph's description.

The parsing of these data files is in textdata.rb, with post-processing in
preprocess.rb in the class TextDataView for generating a table of contents,
links to references, and the actual contents of the page.